requirements.txt
atproto
python-dotenv
transformers
emoji
pandas
nltk
psycopg2
torch
sqlalchemy